+++ Diagnostics/Assets/Turandot/Screen/Turandot.Screen.ScreenElements.cs
@@ -8,12 +8,14 @@ namespace Turandot.Screen
     {
         public List<CueLayout> Cues { get; set; }
         public List<InputLayout> Inputs { get; set; }
+        public bool ApplyCustomScreenColor { get; set; }
         public string finalPrompt = "";
 
         public ScreenElements()
         {
             Cues = new List<CueLayout>();
             Inputs = new List<InputLayout>();
+            ApplyCustomScreenColor = false;
         }
     }
 }